Patched a bug for case when input graph is a tree.
Internally, the given graph is reconstructed, and in the reconstruction
the directions of some edges may be reversed. This was causing problems
when we finally attempted to set connector routes in the original graph.
Those edges that had been reversed were unable to be located, since their
description was reversed. Now we check for the reverse description as well.